+++ rust/src/consumers/mod.rs
@@ -20,3 +20,6 @@ pub mod flattening;
 pub mod exp_definable;
 
 pub const TEMPORARY_WIRES_START: u64 = 1u64<<63;
+
+// ir to r1cs converter
+pub mod to_r1cs;