+++ rust/src/producers/mod.rs
@@ -12,5 +12,3 @@ pub mod from_r1cs;
 
 /// gates builder and interface
 pub mod builder;
-// ir to r1cs converter
-pub mod to_r1cs;